+++ tests/test_oml/test_datasets/test_audio.py
@@ -4,8 +4,18 @@ import pandas as pd
 import pytest
 
 from oml.const import PATHS_COLUMN, START_TIME_COLUMN
-from oml.datasets import AudioBaseDataset
+from oml.datasets.audios import (
+    AudioBaseDataset,
+    AudioLabeledDataset,
+    AudioQueryGalleryDataset,
+    AudioQueryGalleryLabeledDataset,
+)
 from oml.utils import get_mock_audios_dataset
+from tests.test_oml.test_datasets.test_datasets import (
+    check_base,
+    check_labeled,
+    check_query_gallery,
+)
 
 
 @pytest.fixture(scope="module")
@@ -52,3 +62,29 @@ def test_start_times(df: pd.DataFrame) -> None:
     for _ in dataset:
         pass
     assert True, "Dataset iteration failed with start times"
+
+
+@pytest.mark.needs_optional_dependency
+def test_audio_datasets() -> None:
+    df_train, df_val = get_mock_audios_dataset(global_paths=True)
+    df = pd.concat([df_train, df_val])
+
+    # Base
+    dataset_b = AudioBaseDataset(paths=df[PATHS_COLUMN].tolist())
+    check_base(dataset_b)
+
+    # Labeled
+    dataset_l = AudioLabeledDataset(df_train)
+    check_base(dataset_l)
+    check_labeled(dataset_l, df_train)
+
+    # Query Gallery
+    dataset_qg = AudioQueryGalleryDataset(df_val)
+    check_base(dataset_qg)
+    check_query_gallery(dataset_qg, df_val)
+
+    # Query Gallery Labeled
+    dataset_qgl = AudioQueryGalleryLabeledDataset(df_val)
+    check_base(dataset_qgl)
+    check_query_gallery(dataset_qgl, df_val)
+    check_labeled(dataset_qgl, df_val)
